### 2025-02-27 #7856 @danielskatz
👋 @rouson - thanks for your submission. I'll edit it.

Can you work on the possibly missing DOI that editorialbot suggests, but note that it may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.

Also, please suggest ~5 potential (non-conflicted) reviewers. You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission. Or people who aren't in the JOSS system would also be ok.  If you know their GitHub usernames, then please mention them here in this thread (without tagging them with an @).
