### 2025-03-04 #7856 @bonachea
Thanks @danielskatz !

> Can you work on the possibly missing DOI that editorialbot suggests, but note that it may be incorrect.

The "missing DOI" is indeed a false positive match (i.e. incorrect).

The bot thinks it found a DOI for reference 6, but it's wrong. The DOI it found actually corresponds to the paper cited as reference 20 (which happens to have the same title), and we provided different DOI for the open-access version of the paper in reference 20.

> please suggest ~5 potential (non-conflicted) reviewers

We'll work on this.
