### 2025-03-04 #7856 @tclune
> 👋 [@tclune](https://github.com/tclune) - Would you be able to review this submission? (assuming you are not conflicted)

@danielskatz I work closely with the author on Fortran language standards efforts and am aware of this work through casual conversation and the occasional informal presentation.   Presuming that none of this constitutes a conflict of interest, I am able to review.

Of course if there is a furlough at the end of next week ....
