### 2025-03-04 71f4897
Update src/caffeine/prif_private_s.F90
Co-authored-by: Katherine Rasmussen <krasmussen@lbl.gov>
