+++ src/caffeine/prif_private_s.F90
@@ -247,7 +247,6 @@ contains
     integer(c_int), intent(in) :: image_num
     integer(c_intptr_t), intent(out) :: ptr
 
-    integer(c_int) :: num_img
 
     call_assert_describe(image_num > 0 .and. image_num <= initial_team%num_images, "base_pointer: image_num not within valid range")
     ptr = caf_convert_base_addr(coarray_handle%info%coarray_data, image_num)
