### 2025-03-05 #7856 @tclune
@danielskatz I believe I can be impartial and meet the criteria in the policy you referenced.  Cheers.
