### 2025-03-06 5957563
CONTRIBUTING: Add a contributors list, with ORCIDs
This fulfills ASCR requirements to add ORCIDs for national lab authors
to software documentation.
