### 2025-03-09 766ae11
Implement prif_{put,get}_strided
These just perform remote address translation and then call the `*_indirect`
variants, as donw with contiguous RMA.

Add some comments to improve readability
