### 2025-03-09 4cd9cbc
Implement prif_{put,get}_strided_indirect
