### 2025-03-10 1b19bea
Add caf_strided_test
