### 2025-03-10 facb295
CI: Always configure GASNet in `--enable-debug` mode
Configuring GASNet in `--enable-debug` mode activates thousands of
sanity-checking assertions inside libgasnet, which helps to detect
usage violations from calls made by Caffeine.

Also enable fpm test verbosity to log compile commands.
