+++ .github/workflows/build-with-flang.yml
@@ -14,6 +14,7 @@ jobs:
     env:
       FC: flang-new
       CC: clang
+      GASNET_CONFIGURE_ARGS: --enable-rpath --enable-debug
 
     steps:
     - name: Checkout code
@@ -38,4 +39,4 @@ jobs:
     - name: Run unit tests
       run: |
         export GASNET_PSHM_NODES=8
-        ./build/run-fpm.sh test -- -d
+        ./build/run-fpm.sh test --verbose -- -d
