### 2025-03-10 736c0b7
CI: Always use the latest setup-fpm
Because we are also defaulting to always use the latest fpm
