+++ .github/workflows/build-with-gfortran.yml
@@ -17,7 +17,7 @@ jobs:
     - name: Checkout code
       uses: actions/checkout@v1
 
-    - uses: fortran-lang/setup-fpm@v6.0.1
+    - uses: fortran-lang/setup-fpm@main
       with:
         github-token: ${{ secrets.GITHUB_TOKEN }}
 
