### 2025-03-10 0b044df
CI: print tool versions immediately before install
