+++ .github/workflows/build-with-flang.yml
@@ -27,13 +27,15 @@ jobs:
     - name: Build with LLVM Flang
       if: contains(matrix.os, 'ubuntu')
       run: |
+        set -x
+        apt update
+        apt install -y build-essential pkg-config make
+        echo == TOOL VERSIONS ==
         fpm --version
         $FC --version
         $CC --version
         export FPM_FC=$FC
         export FPM_CC=$CC
-        apt update
-        apt install -y build-essential pkg-config make
         ./install.sh
 
     - name: Run unit tests
