+++ .github/workflows/build-with-gfortran.yml
@@ -24,13 +24,25 @@ jobs:
     - name: Install on Ubuntu
       if: contains(matrix.os, 'ubuntu')
       run: |
+        set -x
         sudo apt update
         sudo apt install -y build-essential gfortran-14 g++-14 pkg-config make
+        echo == TOOL VERSIONS ==
+        export GCC_VERSION=14
+        fpm --version
+        gfortran-${GCC_VERSION} --version
+        gcc-${GCC_VERSION} --version
         ./install.sh --prefix=${PREFIX}
 
     - name: Install on macOS
       if: contains(matrix.os, 'macos')
       run: |
+        set -x
+        echo == TOOL VERSIONS ==
+        export GCC_VERSION=14
+        fpm --version
+        gfortran-${GCC_VERSION} --version
+        gcc-${GCC_VERSION} --version
         ./install.sh --prefix=${PREFIX}
 
     - name: Run unit tests
