### 2025-03-10 fc2ebae
CI: Ensure job names are self-contained
