+++ .github/workflows/build-with-flang.yml
@@ -4,7 +4,7 @@ on: [push, pull_request]
 
 
 jobs:
-  Build:
+  flang-build:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
