### 2025-03-10 c2ce47a
Implement prif_{failed,stopped}_images, prif_image_status
Provide trivial but nevertheless compliant implementations.

This is all the stopped/failed image support Caffeine is likely to
provide in the near-term anyhow.
