### 2025-03-10 0d2ec0d
Test prif_{failed,stopped}_images, prif_image_status
