### 2025-03-10 5b36459
docs/implementation-status.md: Update Image Queries
