### 2025-03-10 d3f80f8
Implement prif_size_bytes
Fixes #113.
