### 2025-03-11 5c34730
Apply suggestions from code review
Co-authored-by: Damian Rouson <rouson@lbl.gov>
