### 2025-03-11 fb6fa2d
Implement `prif_sync_team`, `prif_sync_memory`
Rename the misleadingly named synchronization_s submodule to match
the clearer terminology in the PRIF spec.
