### 2025-03-11 84447b7
docs/implementation-status.md: Update SYNC Statements
