### 2025-03-11 6160c96
Implement `prif_[lu]cobound*` and `prif_coshape`
