### 2025-03-11 aaf683f
docs/implementation-status.md: Update Coarray Queries
