+++ docs/implementation-status.md
@@ -86,9 +86,9 @@ in the following sections.
 |-----------|--------|-------|
 | `prif_set_context_data`, `prif_get_context_data` | no |  |
 | `prif_size_bytes`                                | **YES** |  |
-| `prif_lcobound_no_dim`, `prif_lcobound_with_dim` | no |  |
-| `prif_ucobound_no_dim`, `prif_ucobound_with_dim` | no |  |
-| `prif_coshape`                                   | no |  |
+| `prif_lcobound_no_dim`, `prif_lcobound_with_dim` | **YES** |  |
+| `prif_ucobound_no_dim`, `prif_ucobound_with_dim` | **YES** |  |
+| `prif_coshape`                                   | **YES** |  |
 | `prif_local_data_pointer`                        | **YES** |  |
 | `prif_image_index`                               | **YES** |  |
 | `prif_image_index_with_team`                     | no |  |
@@ -96,7 +96,7 @@ in the following sections.
 
 ---
 
-## Continguous Coarray Access
+## Contiguous Coarray Access
 ### Support = partial
 
 | Procedure | Status | Notes |
