### 2025-03-12 da5ba16
Move prif_queries_s into coarray_queries_s, to better match PRIF spec
There are no changes in this commit, just textual code movement
