### 2025-03-12 9975713
Relocate coarray_handle_check into prif_private_s
Add sanity-check invocations from more places
