### 2025-03-12 e0fe114
test/main: Re-enable stop test for flang
