### 2025-03-12 523e734
test/main: Re-enable error stop test for flang
