### 2025-03-12 459175c
Add CMDMSG to asserts in stop tests
When execute_command_line fails, the cmdmsg often provides more
information about what went wrong, so always include it as the
optional message in assertions.
