+++ test/prif_stop_test.F90
@@ -6,6 +6,8 @@ module caf_stop_test
     private
     public :: test_prif_this_image
 
+   integer, parameter :: max_message_len = 128
+
 contains
     function test_prif_this_image() result(tests)
         type(test_item_t) :: tests
@@ -22,57 +24,71 @@ contains
     function exit_with_no_stop_code() result(result_)
         type(result_t) :: result_
         integer exit_status, cmd_stat
+        character(len=max_message_len) command_message
+
+        command_message = "exit_with_no_stop_code"
 
         call execute_command_line( &
           command = "./build/run-fpm.sh run --example stop_with_no_code > /dev/null 2>&1", &
           wait = .true., &
           exitstat = exit_status, &
-          cmdstat = cmd_stat &
+          cmdstat = cmd_stat, &
+          cmdmsg = command_message &
         )
-        result_ = assert_equals(0, exit_status)
+        result_ = assert_equals(0, exit_status, command_message)
 
     end function
 
     function exit_with_integer_stop_code() result(result_)
         type(result_t) :: result_
         integer exit_status, cmd_stat
+        character(len=max_message_len) command_message
+
+        command_message = "exit_with_integer_stop_code"
 
         call execute_command_line( &
           command = "./build/run-fpm.sh run --example stop_with_integer_code > /dev/null 2>&1", &
           wait = .true., &
           exitstat = exit_status, &
-          cmdstat = cmd_stat &
+          cmdstat = cmd_stat, &
+          cmdmsg = command_message &
         )
-        result_ = assert_equals(expected_stop_code, exit_status)
+        result_ = assert_equals(expected_stop_code, exit_status, command_message)
 
     end function
 
     function exit_with_character_stop_code() result(result_)
         type(result_t) :: result_
         integer exit_status, cmd_stat
+        character(len=max_message_len) command_message
+
+        command_message = "exit_with_character_stop_code"
 
         call execute_command_line( &
           command = "./build/run-fpm.sh run --example stop_with_character_code > /dev/null 2>&1", &
           wait = .true., &
           exitstat = exit_status, &
-          cmdstat = cmd_stat &
+          cmdstat = cmd_stat, &
+          cmdmsg = command_message &
         )   
-        result_ = assert_equals(0, exit_status) ! the standard recommends zero exit status for character stop codes
+        ! the standard recommends zero exit status for character stop codes
+        result_ = assert_equals(0, exit_status, command_message) 
 
     end function
 
     function check_callback_invocation() result(result_)
-      type(result_t) :: result_
+        type(result_t) :: result_
+        integer exit_status, cmd_stat
+        character(len=max_message_len) command_message
 
-      integer exit_status, cmd_stat
-
-      call execute_command_line( &
-        command = "./build/run-fpm.sh run --example register_stop_callback > /dev/null 2>&1", &
-        wait = .true., &
-        exitstat = exit_status, &
-        cmdstat = cmd_stat &
-      )
-      result_ = assert_equals(0, exit_status)
+        call execute_command_line( &
+          command = "./build/run-fpm.sh run --example register_stop_callback > /dev/null 2>&1", &
+          wait = .true., &
+          exitstat = exit_status, &
+          cmdstat = cmd_stat, &
+          cmdmsg = command_message &
+        )
+        result_ = assert_equals(0, exit_status, command_message)
     end function
 
 end module caf_stop_test
