### 2025-03-12 597939a
Output flang version as part of test warning header
Also flush output at key points to increase the chances it arrives
at the console in the case of a job crash.
