### 2025-03-12 b2f789f
Fix inadvertent parallelism in (error) stop tests
Previously the stop and error stop tests had EVERY image in an N-image job
launching a recursive `fpm run` invocation that leads to an N-way exit test,
for a total of (N + N^2) concurrent processes, where only (N + N) was intended.

Change the stop tests so that only the first image performs the recursive fpm
invocation for the sub-job.

Partially addresses #137
