+++ test/prif_stop_test.F90
@@ -1,15 +1,16 @@
 module caf_stop_test
-    use veggies, only: test_item_t, describe, result_t, it, assert_that, assert_equals
+    use prif, only: prif_this_image_no_coarray, prif_sync_all
+    use veggies, only: test_item_t, describe, result_t, it, assert_that, assert_equals, succeed
     use unit_test_parameters_m, only : expected_stop_code
 
     implicit none
     private
-    public :: test_prif_this_image
+    public :: test_prif_stop
 
    integer, parameter :: max_message_len = 128
 
 contains
-    function test_prif_this_image() result(tests)
+    function test_prif_stop() result(tests)
         type(test_item_t) :: tests
 
         tests = describe( &
@@ -21,11 +22,21 @@ contains
                 ])
     end function
 
+    function image_one() result(result_)
+        logical :: result_
+        integer :: me
+
+        call prif_sync_all()
+        call prif_this_image_no_coarray(this_image=me)
+        result_ = (me == 1)
+    end function 
+
     function exit_with_no_stop_code() result(result_)
         type(result_t) :: result_
         integer exit_status, cmd_stat
         character(len=max_message_len) command_message
 
+      if (image_one()) then
         command_message = "exit_with_no_stop_code"
 
         call execute_command_line( &
@@ -36,6 +47,9 @@ contains
           cmdmsg = command_message &
         )
         result_ = assert_equals(0, exit_status, command_message)
+      else
+        result_ = succeed("skipped")
+      end if
 
     end function
 
@@ -44,6 +58,7 @@ contains
         integer exit_status, cmd_stat
         character(len=max_message_len) command_message
 
+      if (image_one()) then
         command_message = "exit_with_integer_stop_code"
 
         call execute_command_line( &
@@ -54,6 +69,9 @@ contains
           cmdmsg = command_message &
         )
         result_ = assert_equals(expected_stop_code, exit_status, command_message)
+      else
+        result_ = succeed("skipped")
+      end if
 
     end function
 
@@ -62,6 +80,7 @@ contains
         integer exit_status, cmd_stat
         character(len=max_message_len) command_message
 
+      if (image_one()) then
         command_message = "exit_with_character_stop_code"
 
         call execute_command_line( &
@@ -73,6 +92,9 @@ contains
         )   
         ! the standard recommends zero exit status for character stop codes
         result_ = assert_equals(0, exit_status, command_message) 
+      else
+        result_ = succeed("skipped")
+      end if
 
     end function
 
@@ -81,6 +103,7 @@ contains
         integer exit_status, cmd_stat
         character(len=max_message_len) command_message
 
+      if (image_one()) then
         call execute_command_line( &
           command = "./build/run-fpm.sh run --example register_stop_callback > /dev/null 2>&1", &
           wait = .true., &
@@ -89,6 +112,9 @@ contains
           cmdmsg = command_message &
         )
         result_ = assert_equals(0, exit_status, command_message)
+      else
+        result_ = succeed("skipped")
+      end if
     end function
 
 end module caf_stop_test
