### 2025-03-12 76191d4
stop tests: add an envvar knob to control sub-job parallelism
Add a "SUBJOB_PREFIX" envvar knob that can be used to control the
sub-job launch command in the exit tests
