+++ test/prif_error_stop_test.F90
@@ -1,7 +1,8 @@
 module caf_error_stop_test
     use prif, only: prif_this_image_no_coarray, prif_sync_all
     use veggies, only: test_item_t, describe, result_t, it, assert_that, assert_equals, succeed
-    use unit_test_parameters_m, only : expected_error_stop_code
+    use unit_test_parameters_m, only : expected_error_stop_code, &
+        image_one => subjob_setup, cmd_prefix => subjob_prefix
 
     implicit none
     private
@@ -21,15 +22,6 @@ contains
                 ])
     end function
 
-    function image_one() result(result_)
-        logical :: result_
-        integer :: me
-
-        call prif_sync_all()
-        call prif_this_image_no_coarray(this_image=me)
-        result_ = (me == 1)
-    end function 
-
     function exit_with_no_stop_code() result(result_)
         type(result_t) :: result_
         integer exit_status
@@ -40,7 +32,7 @@ contains
         command_message = "exit_with_no_stop_code"
 
         call execute_command_line( &
-          command = "./build/run-fpm.sh run --example error_stop_with_no_code > /dev/null 2>&1" &
+          command = cmd_prefix//"./build/run-fpm.sh run --example error_stop_with_no_code > /dev/null 2>&1" &
          ,wait = .true. &
          ,exitstat = exit_status &
          ,cmdstat = command_status &
@@ -63,7 +55,7 @@ contains
         command_message = "exit_with_integer_stop_code"
 
         call execute_command_line( &
-          command = "./build/run-fpm.sh run --example error_stop_with_integer_code > /dev/null 2>&1" &
+          command = cmd_prefix//"./build/run-fpm.sh run --example error_stop_with_integer_code > /dev/null 2>&1" &
          ,wait = .true. &
          ,exitstat = exit_status &
          ,cmdstat = command_status &
@@ -87,7 +79,7 @@ contains
         command_message = "exit_with_character_stop_code"
 
         call execute_command_line( &
-          command = "./build/run-fpm.sh run --example error_stop_with_character_code > /dev/null 2>&1" &
+          command = cmd_prefix//"./build/run-fpm.sh run --example error_stop_with_character_code > /dev/null 2>&1" &
          ,wait = .true. &
          ,exitstat = exit_status &
          ,cmdstat = command_status &
