### 2025-03-13 #7895 @editorialbot
[REVIEW]: Caffeine: A parallel runtime library for supporting Fortran compilers
**Submitting author:** @rouson (<a href="http://orcid.org/0000-0002-2344-868X">Damian Rouson</a>)
**Repository:** https://github.com/BerkeleyLab/caffeine
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** 0.5.0
