### 2025-03-13 #7895 @tclune
@danielskatz There is the possibility of a furlough which might impact my ability to meet the 2-4 week goal.
