### 2025-03-14 9c06ccd
Apply suggestions from code review
use an associate block in coarray_handle_check

Co-authored-by: Katherine Rasmussen <krasmussen@lbl.gov>
