### 2025-03-14 4c9b2b3
Apply suggestions from code review
Use an associate block in prif_coshape

Co-authored-by: Katherine Rasmussen <krasmussen@lbl.gov>
