### 2025-03-14 7814d6a
Make coarray_handle_check elemental
As suggested by Kate
