### 2025-03-14 8bb8199
Tweak coarray_handle_check
Change coarray_handle_check slightly to unconditionally validate
invariants.

Add comment documentation.
