+++ src/image_queries_s.F90
@@ -1,5 +1,8 @@
 ! Copyright (c), The Regents of the University of California
 ! Terms of use are as specified in LICENSE.txt
+
+#include "assert_macros.h"
+
 submodule(prif:prif_private_s) image_queries_s
 
   implicit none
@@ -27,23 +30,30 @@ contains
   end procedure
 
   module procedure prif_this_image_with_coarray
+    call_assert(coarray_handle_check(coarray_handle))
+
     call unimplemented("prif_this_image_with_coarray")
   end procedure
 
   module procedure prif_this_image_with_dim
+    call_assert(coarray_handle_check(coarray_handle))
+
     call unimplemented("prif_this_image_with_dim")
   end procedure
 
   module procedure prif_failed_images
-    call unimplemented("prif_failed_images")
+    ! no current support for detecting image failure
+    allocate(failed_images(0))
   end procedure
 
   module procedure prif_stopped_images
-    call unimplemented("prif_stopped_images")
+    ! no current support for detecting image stops
+    allocate(stopped_images(0))
   end procedure
 
   module procedure prif_image_status
-    call unimplemented("prif_image_status")
+    ! no current support for detecting image failure/stops
+    image_status = 0
   end procedure
 
 end submodule image_queries_s
