### 2025-03-15 BerkeleyLab/caffeine#188 @MichaelSiehl
Caffeine installation process using install.sh
I am using ```install.sh``` actually to install CAFFEINE as part of the JOSS review process (openjournals/joss-reviews#7895) using a (fresh) Ubuntu 24.04.2 LTS system. Here are my findings:

As already explained with CAFFEINE issue https://github.com/BerkeleyLab/caffeine/issues/97 using CAFFEINE ```install.sh``` to install **fpm** does not work. The issue with installing **fpm** is still there, problem description and solution can be found here:
https://fortran-lang.discourse.group/t/installing-fpm-0-10-1/8268/5

I did use the following to install **fpm** successfully manually:
```
git clone https://github.com/fortran-lang/fpm.git
cd fpm
sh ./install.sh
```

I did also install **Homebrew** manually because CAFFEINE ```install.sh``` does not install **curl** and **git**.

### Here is my complete successful CAFFEINE installation process:

1. Install **Homebrew** manually as it is explained here: https://linuxconfig.org/installing-homebrew-on-ubuntu-debian-a-step-by-step-guide .

2. Run the CAFFEINE ```install.sh``` to install the compilers.

3. Install **fpm** manually as explained:
```
git clone https://github.com/fortran-lang/fpm.git
cd fpm
sh ./install.sh
```
and add **fpm** to your ```PATH```.

4. Open a new terminal window and re-run the CAFFEINE ```install.sh``` to successfully complete the installation process.
