### 2025-03-17 9ed2ce7
CI: Add run of some examples
This ensures our hello example gets coverage, and also establishes
whether basic Caffeine functionality is working, separately from
the veggies test harness.

Also separately run a few of the stop unit tests to
validate exit code propagation.
