### 2025-03-19 ed6994d
Re-enable co_min and co_max testing for flang
These now pass using our flang-new snapshot on perlmutter:

flang version 21.0.0git (git@github.com:llvm/llvm-project.git 5929de8c7731748bf58ad9b1fedfed75e7aae455)

co_sum test remains broken at runtime.
