+++ test/main.F90
@@ -92,10 +92,10 @@ contains
         individual_tests = [individual_tests, caf_allocate_prif_allocate()]
         individual_tests = [individual_tests, caf_coarray_inquiry_coarray_inquiry()]
         individual_tests = [individual_tests, caf_co_broadcast_prif_co_broadcast()]
-#if !__flang__
         individual_tests = [individual_tests, caf_co_max_prif_co_max()]
         individual_tests = [individual_tests, caf_co_min_prif_co_min()]
         individual_tests = [individual_tests, caf_co_reduce_prif_co_reduce()]
+#if !__flang__ 
         individual_tests = [individual_tests, caf_co_sum_prif_co_sum()]
 #endif
         individual_tests = [individual_tests, caf_image_index_prif_image_index()]
