### 2025-03-19 f956d8b
fix #188: check for git in install.sh
If the install.sh script can't execute git, the script will now
print a message early requesting the user install git, after which
the installer exits with non-zero status.
