### 2025-03-19 0d1402d
caffeine.c: Fix a harmless warning
