### 2025-03-19 20d4dbd
re-enable co_sum test for flang
