+++ test/main.F90
@@ -71,7 +71,7 @@ contains
 
         allocate(individual_tests(0))
 
-#if __flang__
+#if __flang__ && 0 /* currently no disabled tests */
     block
         integer :: major, minor
 #     if defined(__flang_major__) && defined(__flang_minor__)
@@ -95,9 +95,7 @@ contains
         individual_tests = [individual_tests, caf_co_max_prif_co_max()]
         individual_tests = [individual_tests, caf_co_min_prif_co_min()]
         individual_tests = [individual_tests, caf_co_reduce_prif_co_reduce()]
-#if !__flang__ 
         individual_tests = [individual_tests, caf_co_sum_prif_co_sum()]
-#endif
         individual_tests = [individual_tests, caf_image_index_prif_image_index()]
         individual_tests = [individual_tests, caf_num_images_prif_num_images()]
         individual_tests = [individual_tests, test_prif_image_queries()]
