### 2025-03-19 65060c5
fix(install.sh): check earlier for curl
