### 2025-03-19 BerkeleyLab/caffeine#188 @bonachea
Hello Michael, thank you for pointing out that our script doesn’t check for `git` before using `fpm` in a way that will cause `git` to be invoked.

In response to this issue, PR #190 added checks for `git` and `curl` to the install script and with an explanatory message asking users to install the missing dependency if it is not found.

We've also submitted an [issue](https://github.com/fortran-lang/homebrew-fortran/issues/70) to the relevant fortran-lang/homebrew-fortran project so they consider adding `git` as a dependency in their `fpm` Homebrew recipe.
