### 2025-03-24 #7895 @tclune
@danielskatz I feel like a broken record, but somehow I always struggle to find the actual paper with these submissions.  From what I can read above there should be a "paper.md" on the joss-paper branch of the repository.    I did not find any such file there.    Apologies for the trouble.
