### 2025-03-24 #7895 @danielskatz
@tclune - generally, you can scroll up in the issue to find the pdf of the paper, in this case in https://github.com/openjournals/joss-reviews/issues/7895#issuecomment-2721540666

Also, to get to the source, the first comment in the issue is
> Repository: https://github.com/BerkeleyLab/caffeine
Branch with paper.md (empty if default branch): joss-paper

which tells you the repo and branch. It's then up to the authors where the paper source is, but in this case, it's in `docs/joss`, or more directly on the web, https://github.com/BerkeleyLab/caffeine/tree/joss-paper/docs/joss
