### 2025-03-24 f21e115
Re-enable co_{min,max}_character implementations
Fix a critical bug in char_max_wrapper

Fixes #151
Fixes #152
