### 2025-03-24 b24b1b2
co_reduce: Deploy bug workaround for gfortran
Be more explicit and add more sanity-check assertions
