+++ src/caffeine/caffeine.c
@@ -206,8 +206,17 @@ void caf_sync_team( gex_TM_t team ) {
 
 void caf_co_reduce(
   CFI_cdesc_t* a_desc, int result_image, size_t num_elements, gex_Coll_ReduceFn_t user_op, void* client_data, gex_TM_t team
-)
-{
+) {
+  assert(a_desc);
+  assert(result_image >= 0);
+  assert(num_elements > 0);
+  assert(user_op);
+#if PLATFORM_COMPILER_GNU
+  // gfortran 13.2 c_funloc is non-compliant
+  // it erroneously generates a non-callable pointer to a pointer to the subroutine
+  // Here we undo that incorrect extra level of indirection
+  user_op = *(gex_Coll_ReduceFn_t *)user_op; 
+#endif
   char* a_address = (char*) a_desc->base_addr;
   size_t c_sizeof_a = a_desc->elem_len;
   gex_Event_t ev;
