### 2025-03-24 90e7b4b
Re-enable tests for prif_co_{min,max}_character
