### 2025-03-24 f5ad27e
Re-enable tests for prif_co_{min,max}_character
