### 2025-03-25 773ce85
Remove automated invocation of FORD
The PRIF specification: https://go.lbl.gov/prif
is the best and only reliable documentation for clients of the
PRIF interface exported by Caffeine.

The documentation generated by FORD from the repo lacks all semantic
information, and provides no clear-cut distinction between what is part
of the stable PRIF interface and what is an internal Caffeine
implementation detail.

As such, the FORD docs do more harm than good, so stop generating them
automatically.
