+++ .github/workflows/deploy-docs.yml
@@ -1,43 +1,0 @@
-name: Build and Deploy Documentation
-
-on: [push, pull_request]
-
-jobs:
-  docs-build:
-    runs-on: ubuntu-22.04
-
-    steps:
-    - name: Checkout code
-      uses: actions/checkout@v4
-
-    - name: Install Dependencies Ubuntu
-      run: |
-        sudo apt-get update
-        sudo apt install -y python3-dev python3 build-essential graphviz
-        sudo python3 -m pip install ford markdown==3.3.4
-
-    - name: Build Developer Documentation
-      run: |
-        ford doc-generator.md
-
-    - name: Upload Documentation
-      uses: actions/upload-artifact@v4
-      with:
-        name: documentation
-        path: doc/html
-        if-no-files-found: error
-
-    - name: Broken Link Check
-      if: ${{ github.ref == 'refs/heads/main'}}
-      uses: technote-space/broken-link-checker-action@v2.3.1
-      with:
-        TARGET: file://${{ github.workspace }}/doc/html/index.html
-        RECURSIVE: true
-        ASSIGNEES: ${{ github.actor }}
-
-    - name: Deploy API Documentation
-      uses: JamesIves/github-pages-deploy-action@v4.6.4
-      if: ${{ github.event_name == 'push'  &&  github.ref == 'refs/heads/main' }}
-      with:
-        branch: gh-pages
-        folder: doc/html
