### 2025-03-25 765d052
FIXUP: (chore): Pre-release chores
Fix a trivial typo
