### 2025-03-25 0a140a6
Fix misc harmless warnings from gfortran-13
Some examples:
```
test/main.F90:91:74:

   91 |         individual_tests = [individual_tests, a00_caffeinate_caffeinate()]
      |                                                                          ^
Warning: 'MEM[(struct test_item_t[1] *)_1380][0]' may be used uninitialized [-Wmaybe-uninitialized]

Warning: Unused parameter 'initiation_success' declared at (1) [-Wunused-parameter]
test/prif_this_image_test.F90:28:58:

   28 |         image_numbers = [(merge(0, me, me/=i), i = 1, ni)]
      |                                                    
[…]
