### 2025-03-25 5431b9a
Fix harmless warning from gfortran-13 in `as_int()`
```
././src/caffeine/prif_private_s.F90:283:47:

  283 |     integer(int_ptr_check), parameter :: unused = 0_int_ptr_check
      |                                               1
Warning: Unused parameter 'unused' declared at (1) [-Wunused-parameter]
```

Use a real (dynamic) assertion instead.
