### 2025-03-25 4877943
issue #196: Stop injecting outdated/broken flang-new options
`-mmlir -allow-assumed-rank` are unnecessary in current versions of flang-21
that we are using/requiring.

`-Ofast` is a deprecated option that causes our build to spew spurious warnings.
More importantly, it appears responsible for ICEs with the code on this branch.

Resolves #196
