### 2025-03-25 0e29552
test/prif_co_reduce_test: re-enable "easy" co_reduce subtests
* Workaround a lack of F18 intrinsic REDUCE() in gfortran (13)
