### 2025-03-26 ecddfa7
test/prif_co_reduce_test: Disable the subtest requiring parameterized derived types
Our test of prif_co_reduce over an array of parameterized derived type appears
to require compiler support that is lacking in both recent flang and gfortran.
Disable it gated by an optional command-line define HAVE_PARAM_DERIVED

Add comments.
