+++ test/prif_co_reduce_test.F90
@@ -12,15 +12,17 @@ module caf_co_reduce_test
     real :: snd
   end type
 
-  ! type :: array(length)
-  !   integer, len :: length = 2
-  !   integer :: elements(length)
-  ! end type
+#if HAVE_PARAM_DERIVED
+  type :: array(length)
+    integer, len :: length = 2
+    integer :: elements(length)
+  end type
 
-  ! type :: reduction_context_data
-  !   type(c_funptr) :: user_op
-  !   integer :: length
-  ! end type
+  type :: reduction_context_data
+    type(c_funptr) :: user_op
+    integer :: length
+  end type
+#endif
 
 contains
 
@@ -31,7 +33,9 @@ contains
       "The prif_co_reduce subroutine", &
       [ it("can be used to implement logical and reduction", check_logical) &
       , it("can be used for reduction on simple derived types", check_derived_type_reduction) &
+#if HAVE_PARAM_DERIVED
       , it("can be used for reduction on derived types with length type parameters", check_type_parameter_reduction) &
+#endif
       ])
   end function
 
@@ -136,73 +140,80 @@ contains
     end do
   end subroutine
 
+#if HAVE_PARAM_DERIVED
+! As of LLVM20, flang does not implement the types used by this test:
+! flang/lib/Lower/ConvertType.cpp:482: not yet implemented: parameterized derived types
+! error: Actual argument associated with TYPE(*) dummy argument 'a=' may not have a parameterized derived type
+
+! Gfortran 14.2 also lacks the type support for this test:
+! Error: Derived type 'pdtarray' at (1) is being used before it is defined
+
   function check_type_parameter_reduction() result(result_)
     type(result_t) :: result_
-    result_ = succeed("skip for now")
-    
-    ! type(array), parameter :: values(*,*) = reshape( &
-    !     [ array([1, 53]), array([3, 47]) &
-    !     , array([5, 43]), array([7, 41]) &
-    !     , array([11, 37]), array([13, 31]) &
-    !     , array([17, 29]), array([19, 23]) &
-    !     ], &
-    !     [2, 4])
-    ! integer :: me, ni, i
-    ! type(array(values%length)), dimension(size(values,1)) :: my_val, expected
-    ! procedure(prif_operation_wrapper_interface), pointer :: op
-    ! type(reduction_context_data), target :: context
+    type(array), parameter :: values(*,*) = reshape( &
+        [ array(elements=[1, 53]), array(elements=[3, 47]) &
+        , array(elements=[5, 43]), array(elements=[7, 41]) &
+        , array(elements=[11, 37]), array(elements=[13, 31]) &
+        , array(elements=[17, 29]), array(elements=[19, 23]) &
+        ], &
+        [2, 4])
+    integer :: me, ni, i
+    type(array(values%length)), dimension(size(values,1)) :: my_val, expected
+    procedure(prif_operation_wrapper_interface), pointer :: op
+    type(reduction_context_data), target :: context
 
-    ! op => array_wrapper
-    ! context%user_op = c_funloc(add_array)
-    ! context%length = values%length
-    ! call prif_this_image_no_coarray(this_image=me)
-    ! call prif_num_images(ni)
+    op => array_wrapper
+    context%user_op = c_funloc(add_array)
+    context%length = values%length
+    call prif_this_image_no_coarray(this_image=me)
+    call prif_num_images(ni)
 
-    ! my_val = values(:, mod(me-1, size(values,2))+1)
-    ! call prif_co_reduce(my_val, op, c_loc(context))
+    my_val = values(:, mod(me-1, size(values,2))+1)
[…]
