### 2025-03-26 450f77c
implementation-status.md: Update collectives status
