+++ docs/implementation-status.md
@@ -199,13 +199,13 @@ in the following sections.
 
 | Procedure | Status | Notes |
 |-----------|--------|-------|
-| `prif_co_broadcast` | **YES** |  |
-| `prif_co_max`       | *partial* | only supports 32-bit and 64-bit numeric types |
-| `prif_co_max_character` | no | procedure not yet added to Caffeine |
-| `prif_co_min`       | *partial* | only supports 32-bit and 64-bit numeric types |
-| `prif_co_min_character` | no | procedure not yet added to Caffeine |
-| `prif_co_sum`       | *partial* | only supports 32-bit and 64-bit numeric types |
-| `prif_co_reduce`    | *partial* | only supports intrinsic types (no support for derived types), interface not yet updated to v0.5 |
+| `prif_co_broadcast`     | **YES** |  |
+| `prif_co_max`           | *partial* | only supports 32-bit and 64-bit numeric types |
+| `prif_co_max_character` | **YES** | |
+| `prif_co_min`           | *partial* | only supports 32-bit and 64-bit numeric types |
+| `prif_co_min_character` | **YES** | |
+| `prif_co_sum`           | *partial* | only supports 32-bit and 64-bit numeric types |
+| `prif_co_reduce`        | **YES** | |
 
 ---
 
