### 2025-03-26 2a5e44c
(chore): Update README with compiler version info
