### 2025-03-26 1f5defe
(chore): Update README with description of build dependencies
