### 2025-03-26 #7895 @bonachea
@tclune Thanks for your response!

Caffeine's "functionality documentation" is the [PRIF interface specification](https://go.lbl.gov/prif), which is cited in the [repo README.md](https://github.com/BerkeleyLab/caffeine/blob/main/README.md#caffeine-is-an-implementation-of-the-parallel-runtime-interface-for-fortran-prif).

The same document is also cited as this reference in the paper:

> Bonachea, D., Rasmussen, K., Richardson, B., & Rouson, D. (2024b). Parallel runtime
> interface for fortran (PRIF) specification (rev. 0.5). Lawrence Berkeley National Laboratory
> (LBNL), Berkeley, CA (United States). https://doi.org/10.25344/S4CG6G
