### 2025-03-26 #7895 @tclune
> Caffeine's "functionality documentation" is the [PRIF interface specification](https://go.lbl.gov/prif), which is cited in the [repo README.md](https://github.com/BerkeleyLab/caffeine/blob/main/README.md#caffeine-is-an-implementation-of-the-parallel-runtime-interface-for-fortran-prif).

Thanks.    My eyes glossed right over that because it is a section heading.    Documentation looks excellent (as expected).
