### 2025-03-26 080d3fa
CI: Factor, combine and expand our GitHub actions
The Build Action now expands to jobs covering:

* gfortran-{13,14} on ubuntu-24.04, macos-13-x86_64, macos-14-arm64
* flang-new (currently v20 git snapshot) on ubuntu-24.04

In addition, the action recipe is now fully factored, simplifying future maintenance.
