+++ .github/workflows/build-with-gfortran.yml
@@ -1,60 +1,0 @@
-name: Build with gfortran
-
-on: [push, pull_request]
-
-jobs:
-  gfortran-build:
-    runs-on: ${{ matrix.os }}
-    strategy:
-      matrix:
-        os: [macos-13, ubuntu-24.04]
-
-    env:
-      PREFIX: install
-      GASNET_CONFIGURE_ARGS: --enable-rpath --enable-debug
-      GASNET_PSHM_NODES: 8
-
-    steps:
-    - name: Checkout code
-      uses: actions/checkout@v1
-
-    - uses: fortran-lang/setup-fpm@main
-      with:
-        github-token: ${{ secrets.GITHUB_TOKEN }}
-
-    - name: Install on Ubuntu
-      if: contains(matrix.os, 'ubuntu')
-      run: |
-        set -x
-        sudo apt update
-        sudo apt install -y build-essential gfortran-14 g++-14 pkg-config make
-        echo == TOOL VERSIONS ==
-        export GCC_VERSION=14
-        fpm --version
-        gfortran-${GCC_VERSION} --version
-        gcc-${GCC_VERSION} --version
-        ./install.sh --prefix=${PREFIX}
-
-    - name: Install on macOS
-      if: contains(matrix.os, 'macos')
-      run: |
-        set -x
-        echo == TOOL VERSIONS ==
-        export GCC_VERSION=14
-        fpm --version
-        gfortran-${GCC_VERSION} --version
-        gcc-${GCC_VERSION} --version
-        ./install.sh --prefix=${PREFIX}
-
-    - name: Run examples
-      run: |
-        set -x
-        ./build/run-fpm.sh run --verbose --example hello
-        ./build/run-fpm.sh run --verbose --example stop_with_no_code
-        ( set +e ; ./build/run-fpm.sh run --verbose --example stop_with_integer_code ; test $? = 99 )
-        ( set +e ; ./build/run-fpm.sh run --verbose --example error_stop_with_integer_code ; test $? = 100 )
-
-    - name: Run unit tests
-      run: |
-        set -x
-        ./build/run-fpm.sh test --verbose -- -d
