### 2025-03-27 d9f4d28
issue #159: Expand prif_co_{min,max,sum} to handle every interoperable 32- or 64-bit integer type
