+++ src/caffeine/caffeine.c
@@ -246,23 +246,82 @@ void caf_co_broadcast(CFI_cdesc_t * a_desc, int source_image, int num_elements, 
   gex_Event_Wait(ev);
 }
 
-void caf_co_max(CFI_cdesc_t* a_desc, int result_image, size_t num_elements, gex_TM_t team)
-{
-  gex_DT_t a_type;
+//-------------------------------------------------------------------
+// Typed computational collective subroutines
+//-------------------------------------------------------------------
 
-  switch (a_desc->type)
-  {
-    case CFI_type_int32_t:          a_type = GEX_DT_I32; break;
-    case CFI_type_int64_t:          a_type = GEX_DT_I64; break;
-    case CFI_type_float:            a_type = GEX_DT_FLT; break;
-    case CFI_type_double:           a_type = GEX_DT_DBL; break;
-    default:
-      gasnett_fatalerror("Unrecognized type: %d", (int)a_desc->type);
+// Convert CFI_cdesc_t to the corresponding GEX reduction data type
+// returns the size of the native type
+static size_t CFI_to_GEX_DT(CFI_type_t cfi_type, gex_DT_t *gex_dt, int *complex_scale) {
+  assert(gex_dt);
+
+  if_pf (complex_scale) *complex_scale = 1;
+
+  switch (cfi_type) {
+    // real cases
+    case CFI_type_float:            *gex_dt = GEX_DT_FLT; return 4;
+    case CFI_type_double:           *gex_dt = GEX_DT_DBL; return 8;
+
+    // complex cases
+    case float_Complex_workaround:  *gex_dt = GEX_DT_FLT; 
+      if (!complex_scale) gasnett_fatalerror("This operation does not support complex types");
+      *complex_scale = 2;
+      return 8;
+    case double_Complex_workaround: *gex_dt = GEX_DT_DBL; 
+      if (!complex_scale) gasnett_fatalerror("This operation does not support complex types");
+      *complex_scale = 2;
+      return 16;
+    // no support for CFI_type_long_double or CFI_type_long_double_Complex
   }
 
-  char* a_address = (char*) a_desc->base_addr;
+  // integer types
+  #define CFI_INT_CASE(cfi_type_constant, c_type) \
+    else if (cfi_type == cfi_type_constant) { \
+      if (sizeof(c_type) == 4) *gex_dt = GEX_DT_I32;  \
+      else if (sizeof(c_type) > 8) \
+         gasnett_fatalerror("Unsupported wide integer type: %d", (int)cfi_type); \
+      else                     *gex_dt = GEX_DT_I64;  \
+      return sizeof(c_type); \
+    }
+  // these must be handled outside the switch because there are duplicates
+  // for the same reason, start with the most likely candidates
+  if (0) ;
+  CFI_INT_CASE(CFI_type_int64_t, int64_t)
+  CFI_INT_CASE(CFI_type_int32_t, int32_t)
+  CFI_INT_CASE(CFI_type_int16_t, int16_t)
+  CFI_INT_CASE(CFI_type_int8_t, int8_t)
+  CFI_INT_CASE(CFI_type_Bool, _Bool)
+  CFI_INT_CASE(CFI_type_char, char)
+  CFI_INT_CASE(CFI_type_signed_char, signed char)
+  CFI_INT_CASE(CFI_type_short, short int)
+  CFI_INT_CASE(CFI_type_int, int)
[…]
