### 2025-03-27 1896426
Expand prif_co_{min,max,sum}_test to use default integer
