+++ test/prif_co_min_test.F90
@@ -13,12 +13,30 @@ contains
 
         tests = describe( &
           "The prif_co_min subroutine computes the minimum value across images for corresponding elements for", &
-          [ it("32 bit integer scalars", check_32_bit_integer) &
+          [ it("a 1D default integer array", check_default_integer) &
+          , it("32 bit integer scalars", check_32_bit_integer) &
           , it("a 1D 64 bit integer array", check_64_bit_integer) &
           , it("a 2D 32 bit real array", check_32_bit_real) &
           , it("a 1D 64 bit real array", check_64_bit_real) &
           , it("a character scalar", check_character) &
           ])
+    end function
+
+    function check_default_integer() result(result_)
+        type(result_t) :: result_
+
+        integer, parameter :: values(*,*) = reshape([1, 19, 5, 13, 11, 7, 17, 3], [2, 4])
+        integer :: me, ni, i
+        integer, dimension(size(values,1)) :: my_val, expected
+
+        call prif_this_image_no_coarray(this_image=me)
+        call prif_num_images(ni)
+
+        my_val = values(:, mod(me-1, size(values,2))+1)
+        call prif_co_min(my_val)
+
+        expected = minval(reshape([(values(:, mod(i-1,size(values,2))+1), i = 1, ni)], [size(values,1),ni]), dim=2)
+        result_ = assert_equals(int(expected), int(my_val))
     end function
 
     function check_32_bit_integer() result(result_)
