### 2025-03-27 bc06291
Expand prif_co_{min,max,sum}_test to use default integer
