### 2025-03-27 7e5d2de
Expand prif_co_{min,max,sum}_test to use 8- and 16-bit integers
