### 2025-03-27 b8fedd7
Expand prif_co_{min,max,sum}_test to use 8- and 16-bit integers
