+++ test/prif_co_max_test.F90
@@ -1,5 +1,5 @@
 module caf_co_max_test
-    use iso_c_binding, only: c_int32_t, c_int64_t, c_float, c_double
+    use iso_c_binding, only: c_int8_t, c_int16_t, c_int32_t, c_int64_t, c_float, c_double
     use prif, only : prif_co_max, prif_co_max_character, prif_this_image_no_coarray, prif_num_images
     use veggies, only: result_t, test_item_t, assert_equals, describe, it, succeed
 
@@ -14,10 +14,12 @@ contains
         tests = describe( &
           "The prif_co_max subroutine computes the maximum value across images for corresponding elements for", &
           [ it("a 1D default integer array", check_default_integer) &
-          , it("32 bit integer scalars", check_32_bit_integer) &
-          , it("a 1D 64 bit integer array", check_64_bit_integer) &
-          , it("a 2D 32 bit real array", check_32_bit_real) &
-          , it("a 1D 64 bit real array", check_64_bit_real) &
+          , it("a 1D 8-bit integer array", check_8_bit_integer) &
+          , it("a 1D 16-bit integer array", check_16_bit_integer) &
+          , it("32-bit integer scalars", check_32_bit_integer) &
+          , it("a 1D 64-bit integer array", check_64_bit_integer) &
+          , it("a 2D 32-bit real array", check_32_bit_real) &
+          , it("a 1D 64-bit real array", check_64_bit_real) &
           , it("a character scalar", check_character) &
           ])
     end function
@@ -25,9 +27,43 @@ contains
     function check_default_integer() result(result_)
         type(result_t) :: result_
 
-        integer, parameter :: values(*,*) = reshape([1, 19, 5, 13, 11, 7, 17, 3], [2, 4])
+        integer, parameter :: values(*,*) = reshape([1, -19, 5, 13, 11, 7, 17, 3], [2, 4])
         integer :: me, ni, i
         integer, dimension(size(values,1)) :: my_val, expected
+
+        call prif_this_image_no_coarray(this_image=me)
+        call prif_num_images(ni)
+
+        my_val = values(:, mod(me-1, size(values,2))+1)
+        call prif_co_max(my_val)
+
+        expected = maxval(reshape([(values(:, mod(i-1,size(values,2))+1), i = 1, ni)], [size(values,1),ni]), dim=2)
+        result_ = assert_equals(int(expected), int(my_val))
+    end function
+
+    function check_8_bit_integer() result(result_)
+        type(result_t) :: result_
+
+        integer(c_int8_t), parameter :: values(*,*) = reshape(int([1, -19, 5, 13, 11, 7, 17, 3],c_int8_t), [2, 4])
+        integer :: me, ni, i
+        integer(c_int8_t), dimension(size(values,1)) :: my_val, expected
+
+        call prif_this_image_no_coarray(this_image=me)
+        call prif_num_images(ni)
+
+        my_val = values(:, mod(me-1, size(values,2))+1)
+        call prif_co_max(my_val)
+
+        expected = maxval(reshape([(values(:, mod(i-1,size(values,2))+1), i = 1, ni)], [size(values,1),ni]), dim=2)
+        result_ = assert_equals(int(expected), int(my_val))
+    end function
+
+    function check_16_bit_integer() result(result_)
+        type(result_t) :: result_
+
+        integer(c_int16_t), parameter :: values(*,*) = reshape(int([1, -19, 5, 13, 11, 7, 17, 3],c_int16_t), [2, 4])
+        integer :: me, ni, i
+        integer(c_int16_t), dimension(size(values,1)) :: my_val, expected
 
[…]
