### 2025-03-27 6c28ff0
issue #159: Implement prif_co_{min,max,sum} for 8- and 16-bit integer types
Perform a widening conversion into a temporary before the reduction,
and narrow back on the result.

Fixes issue #159.
