### 2025-03-27 c14e112
minor changes in paper.bib
