+++ docs/joss/paper.bib
@@ -1,6 +1,6 @@
 @misc{nag-site,
   author={{Numerical Algorithms Group Ltd}},
-  title = {{nAG Fortran Compiler}},
+  title = {{nAG Fortran} Compiler},
   journal = {Web page},
   year = 2025,
   url = {https://nag.com/fortran-compiler/}
@@ -16,7 +16,7 @@
 
 @misc{ifx-site,
   author={{Intel Corporation}},
-  title = {{Intel Fortran Compiler}},
+  title = {Intel {F}ortran Compiler},
   journal = {Web page},
   year = 2025,
   url = {https://www.intel.com/content/www/us/en/developer/tools/oneapi/fortran-compiler.html}
@@ -24,7 +24,7 @@
 
 @misc{cce-site,
   author={{Hewlett Packard Enterprise}},
-  title = {{Cray Compiling Environment Fortran compiler}},
+  title = {Cray Compiling Environment {F}ortran compiler}},
   journal = {Web page},
   year = 2025,
   url = {https://cpe.ext.hpe.com/docs/latest/getting_started/CPE-CCE-Fortran.html}
@@ -40,7 +40,7 @@
 
 @inproceedings{rice-caf-compiler,
   author = {Y. Dotsenko and C. Coarfa and J. Mellor-Crummey},
-  title = {{A Multi-platform Co-Array Fortran Compiler}},
+  title = {A Multi-platform Co-Array {F}ortran Compiler},
   booktitle = {Proc. 13th International Conference on Parallel Architecture and Compilation Techniques (PACT)},
   year = 2004,
   doi = "10.1109/PACT.2004.1342539",
@@ -67,7 +67,7 @@
 }
 
 @article {numrich1998coarray,
-  title={Co-array fortran for parallel programming},
+  title={Co-array {F}ortran for parallel programming},
   author={Numrich, R. W. and Reid, J. K.},
   year={1998},
   journal={ACM Fortran Forum},
@@ -77,7 +77,7 @@
   doi = {10.1145/289918.289920}
 }
 @inproceedings{rouson2022caffeine,
-  title = {{Caffeine: CoArray Fortran Framework of Efficient Interfaces to Network Environments}},
+  title = {Caffeine: {CoArray Fortran} Framework of Efficient Interfaces to Network Environments},
   author={Rouson, Damian and Bonachea, Dan},
   booktitle={2022 IEEE/ACM Eighth Workshop on the LLVM Compiler Infrastructure in HPC (LLVM-HPC)},
   pages={34--42},
@@ -87,7 +87,7 @@
 }
 
 @inproceedings{bonachea2024prif,
-  title={{Parallel Runtime Interface for Fortran (PRIF): A Multi-Image Solution for LLVM Flang}},
+  title={Parallel Runtime Interface for {Fortran (PRIF)}: A Multi-Image Solution for {LLVM Flang}},
   author={Bonachea, Dan and Rasmussen, Katherine and Richardson, Brad and Rouson, Damian},
   booktitle={Tenth Workshop on the LLVM Compiler Infrastructure in HPC (LLVM-HPC2024)},
   organization={IEEE},
@@ -98,7 +98,7 @@
 
 @techreport{prif-0.5,
   author       = {Bonachea, Dan and Rasmussen, Katherine and Richardson, Brad and Rouson, Damian},
-  title        = {Parallel Runtime Interface for Fortran (PRIF) Specification (Rev. 0.5)},
+  title        = {Parallel Runtime Interface for {Fortran (PRIF)} Specification (Rev. 0.5)},
   institution  = {Lawrence Berkeley National Laboratory (LBNL), Berkeley, CA (United States)},
   doi          = {10.25344/S4CG6G},
   url          = {https://www.osti.gov/biblio/2481719},
@@ -120,7 +120,7 @@
 
[…]
