### 2025-03-27 8a5c05f
minor changes in paper.md
