### 2025-03-27 #7895 @danielskatz
@rouson - I've suggested small changes in https://github.com/BerkeleyLab/caffeine/pull/200 - please merge this, or let me know what you disagree with, then we can continue to publication
