### 2025-03-27 669b109
install.sh: Add a --verbose option
To provide more verbose information about the install process
