### 2025-03-27 bd97014
CI: enable install.sh --verbose
