### 2025-03-27 ee324ca
Default to gfortran -Wno-unused-dummy-argument
Because it currently generates a bunch of spurious warnings that
are basically never helpful
