+++ install.sh
@@ -419,7 +419,7 @@ compiler_version=$($FPM_FC --version)
 if [[ $compiler_version == *llvm* ]]; then
   compiler_flag="-mmlir -allow-assumed-rank -g -Ofast"
 else
-  compiler_flag="-g -O3 -ffree-line-length-0"
+  compiler_flag="-g -O3 -ffree-line-length-0 -Wno-unused-dummy-argument"
 fi
 compiler_flag+=" -DASSERT_MULTI_IMAGE -DASSERT_PARALLEL_CALLBACKS"
 
