### 2025-03-27 220be44
Enable Action to build paper on pull-request
