### 2025-03-27 df2063d
Update docs/joss/paper.bib
Co-authored-by: Dan Bonachea <dobonachea@lbl.gov>
