+++ docs/joss/paper.bib
@@ -120,7 +120,7 @@
 
 @inproceedings{gasnet-lcpc18,
   author =       "Dan Bonachea and Paul H. Hargrove",
-  title =        {GASNet-EX: A High-Performance, Portable Communication Library for Exascale},
+  title =        {{GASNet-EX}: A High-Performance, Portable Communication Library for Exascale},
   month =        "October",
   year =         2018,
   doi =          {10.25344/S4QP4W},
