### 2025-03-27 5a163f9
Update docs/joss/paper.bib
Co-authored-by: Dan Bonachea <dobonachea@lbl.gov>
