+++ docs/joss/paper.bib
@@ -24,7 +24,7 @@
 
 @misc{cce-site,
   author={{Hewlett Packard Enterprise}},
-  title = {Cray Compiling Environment {F}ortran compiler}},
+  title = {{Cray Compiling Environment Fortran compiler}},
   journal = {Web page},
   year = 2025,
   url = {https://cpe.ext.hpe.com/docs/latest/getting_started/CPE-CCE-Fortran.html}
