### 2025-03-27 f1e91e3
Update docs/joss/paper.bib
Co-authored-by: Dan Bonachea <dobonachea@lbl.gov>
