+++ docs/joss/paper.bib
@@ -16,7 +16,7 @@
 
 @misc{ifx-site,
   author={{Intel Corporation}},
-  title = {Intel {F}ortran Compiler},
+  title = {{Intel Fortran Compiler}},
   journal = {Web page},
   year = 2025,
   url = {https://www.intel.com/content/www/us/en/developer/tools/oneapi/fortran-compiler.html}
