### 2025-03-27 e94fb15
Update docs/joss/paper.bib
Co-authored-by: Dan Bonachea <dobonachea@lbl.gov>
