+++ docs/joss/paper.bib
@@ -173,7 +173,7 @@
 
 @misc{class-site,
   author={{Lawrence Berkeley National Laboratory}},
-  title = {Computer Languages & Systems Software Group},
+  title = {{Computer Languages & Systems Software Group}},
   journal = {Web page},
   year =         2025,
   url = {https://go.lbl.gov/class}
