### 2025-03-27 958a690
CI: Update flang container to use real flang releases
