### 2025-03-27 d064d96
Implement prif_{set,get}_context_data
The context data is stored in a reserved field of the descriptor
constructed during the call to prif_allocate_coarray.
