+++ src/caffeine/allocation_s.F90
@@ -54,6 +54,7 @@ contains
     coarray_handle%info%lcobounds(1:size(lcobounds)) = lcobounds
     coarray_handle%info%ucobounds(1:size(ucobounds)) = ucobounds
     call add_to_team_list(coarray_handle)
+    coarray_handle%info%p_context_data = c_loc(coarray_handle%info%reserved)
 
     allocated_memory = coarray_handle%info%coarray_data
     if (caf_have_child_teams()) then
