### 2025-03-27 e771220
implementation-status.md: Update status
Remove outdated partial parentheticals
