### 2025-03-28 #7895 @bonachea
> * Please list the DOI of the archived version here.

@danielskatz Here is the requested DOI, for version 0.5.0: https://doi.org/10.5281/zenodo.15099022
