### 2025-03-28 6e83a81
Test prif_alias_{create,destroy}
