### 2025-03-28 de82571
prif_image_index_test: Fix latent defects exposed by new assertion
Test code was violating precondition of prif_allocate_coarray():

  product(ucobounds - lcobounds + 1) >= num_images()

This basically implies that any call to prif_allocate_coarray()
that does not use the result of prif_num_images to compute the
ucobounds is erroneous.
