### 2025-03-28 #7895 @danielskatz
@bonachea - please check the final proof that this will generate, and confirm it's good (which I will also do).
