### 2025-03-28 #7895 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
