+++ docs/joss/paper.bib
@@ -1,7 +1,7 @@
 @misc{nag-site,
   author={{Numerical Algorithms Group Ltd}},
   title = {{nAG Fortran} Compiler},
-  journal = {Web page},
+  journal = {web page},
   year = 2025,
   url = {https://nag.com/fortran-compiler/}
 }
@@ -9,7 +9,7 @@
 @misc{g95-site,
   author={Andy Vaught},
   title = {{G95} Compiler},
-  journal = {Web page},
+  journal = {web page},
   year = 2013,
   url = {https://g95.sourceforge.net}
 }
@@ -17,7 +17,7 @@
 @misc{ifx-site,
   author={{Intel Corporation}},
   title = {{Intel Fortran Compiler}},
-  journal = {Web page},
+  journal = {web page},
   year = 2025,
   url = {https://www.intel.com/content/www/us/en/developer/tools/oneapi/fortran-compiler.html}
 }
@@ -25,7 +25,7 @@
 @misc{cce-site,
   author={{Hewlett Packard Enterprise}},
   title = {{Cray Compiling Environment Fortran compiler}},
-  journal = {Web page},
+  journal = {web page},
   year = 2025,
   url = {https://cpe.ext.hpe.com/docs/latest/getting_started/CPE-CCE-Fortran.html}
 }
@@ -33,7 +33,7 @@
 @misc{caf-2-site,
   author={Mellor-Crummy, J.},
   title = {{Coarray Fortran 2.0}},
-  journal = {Web page},
+  journal = {web page},
   year = 2011,
   url = {http://caf.rice.edu}
 }
@@ -109,7 +109,7 @@
 
 @inproceedings{fanfarillo2014opencoarrays,
  author = {Fanfarillo, Alessandro and Burnus, Tobias and Cardellini, Valeria and Filippone, Salvatore and Nagle, Dan and Rouson, Damian},
- title = {{OpenCoarrays: Open-source Transport Layers Supporting Coarray Fortran Compilers}},
+ title = {Open{C}oarrays: Open-source Transport Layers Supporting Coarray {F}ortran Compilers},
  booktitle = {Partitioned Global Address Space Programming Models (PGAS)},
  year = {2014},
  _url = {http://doi.acm.org/10.1145/2676870.2676876},
