### 2025-03-29 #7895 @bonachea
> 👋 [@bonachea](https://github.com/bonachea) - I found a few things I missed in the bib the first time, in [BerkeleyLab/caffeine#202](https://github.com/BerkeleyLab/caffeine/pull/202)

@danielskatz - Thanks for the additional polish. Approved and merged!
