### 2025-03-29 #7895 @danielskatz
Sorry, this is my fault - I forgot that Crossref is undergoing some scheduled emergency maintenance and is not accepting deposits for a few hours. I'll try again once the service is back up
