### 2025-03-29 #7895 @danielskatz
The paper is in JOSS as https://joss.theoj.org/papers/10.21105/joss.07895

But at least for me, the DOI (https://doi.org/10.21105/joss.07895) doesn't resolve, so I'm going to keep this issue open until it does.
