### 2025-03-29 #7895 @danielskatz
Ah, the DOI now works for me, so I'll close this issue.

Congratulations and thanks again!
